+++ app/Http/Controllers/StudentController.php
@@ -6,6 +6,7 @@ use Inertia\Inertia;
 use App\Models\Classes;
 use App\Models\Student;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Resources\ClassResource;
 use App\Http\Resources\StudentResource;
 use App\Http\Requests\StoreStudentRequest;
@@ -15,6 +16,8 @@ class StudentController extends Controller
 {
     public function index()
     {
+        Gate::authorize('student_access');
+
         $students = Student::paginate(10);
 
         return Inertia::render('Student/Index', [
@@ -24,6 +27,8 @@ class StudentController extends Controller
 
     public function create()
     {
+        Gate::authorize('student_create');
+
         $classes = ClassResource::collection(Classes::all());
 
         return Inertia::render('Student/Create', [
@@ -33,6 +38,8 @@ class StudentController extends Controller
 
     public function store(StoreStudentRequest $request)
     {
+        Gate::authorize('student_create');
+
         Student::create($request->validated());
 
         return redirect()->route('students.index');
@@ -40,6 +47,8 @@ class StudentController extends Controller
 
     public function edit(Student $student)
     {
+        Gate::authorize('student_edit');
+
         $classes = ClassResource::collection(Classes::all());
 
         return Inertia::render('Student/Edit', [
@@ -50,6 +59,8 @@ class StudentController extends Controller
 
     public function update(UpdateStudentRequest $request, Student $student)
     {
+        Gate::authorize('student_edit');
+
         $student->update($request->validated());
 
         return redirect()->route('students.index');
@@ -57,6 +68,8 @@ class StudentController extends Controller
 
     public function destroy(Student $student)
     {
+        Gate::authorize('student_delete');
+
         $student->delete();
 
         return redirect()->route('students.index');
